requirements.txt
torch
torchvision
torchaudio
torchmetrics
peft==0.18.1
accelerate
transformers
pandas
faiss-cpu
einops
fairscale
lark
loguru
matplotlib
open_clip_torch
pillow
prettytable
pytorch-metric-learning
pytorch-lightning==1.9.0
scikit-learn
sentence-transformers
timm
gdown
opencv-python
kornia


